Organized code files into app subfolders for modular structure
